requirements.txt
django
celery
python-dotenv
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-ses
psycopg2-binary
whitenoise
sentry-sdk
wagtail
django_extensions
django_storages
pytz
drf-rw-serializers
django-crum
Redis
# remove it
django-phonenumber-field
arrow
phonenumbers
django-ratelimit